Implemented configuration updates

+++ ukr-generator.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-source ukr-generator.config
+source SIMD_generator.config
+
+CONFIG_PATH=SIMD-arch
 
 sys_arch=$(uname -p)
 
@@ -32,12 +34,33 @@ if [ $UNROLL -ne 0 ]; then
   _UNROLL="--unroll ${UNROLL}"
 fi
 
+#-Set reorder loads option for ukernel generator
+_OP=""
+if [ "$OP" != "NORMAL" ]; then
+  _OP="--op_b ${OP,,}"
+fi
+
+if [ "$DATA" != "FP32" ]; then
+  echo "ERROR: Actually, only is supported Float 32 bits data type."
+  exit -1
+fi
+
+
+if [ ! -f $CONFIG_PATH/$ARCH_CONFIG ]; then
+    echo "ERROR: The Architecture configure doesn't exist. Please, enter a valid filename."
+    exit -1
+fi
+
+VLEN=$(tail -1 $CONFIG_PATH/$ARCH_CONFIG | cut -f1)
+MAXVECTOR=$(tail -1 $CONFIG_PATH/$ARCH_CONFIG | cut -f2)
+
 rm src/asm_generator/ukernels/*
 
 if [ "$ARCH" = "RISCV" ]; then
-  ./src/asm_generator/ukernels_generator.py --arch riscv 
-  #./src/asm_generator/ukernels_generator.py --arch riscv --op_b ${OP,,} ${_PIPELINING} ${_REORDER} ${_UNROLL}
+  ./src/asm_generator/ukernels_generator.py --arch riscv --data ${DATA} --vlen ${VLEN} \
+	                                    --maxvec ${MAXVECTOR} ${_OP} ${_REORDER} ${_PIPELINING} ${_UNROLL}
 else
-  ./src/asm_generator/ukernels_generator.py --arch armv8 ${_PIPELINING} ${_UNROLL}
+  ./src/asm_generator/ukernels_generator.py --arch armv8 --data ${DATA} --vlen ${VLEN} \
+	                                    --maxvec ${MAXVECTOR} ${_PIPELINING} ${_UNROLL}
 fi
